Move realfill folder
